pyproject.toml
[project]
name = "ragsystem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-pinecone>=0.2.11",
    "pathlib2>=2.3.7.post1",
    "pdfplumber>=0.11.7",
    "pinecone>=7.3.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.1",
]
